+++ history/include/history.h
@@ -32,11 +32,11 @@
     /// History ///
     ///////////////
 
-    size_t get_file_nb_lines(char *filename);
+    int get_file_nb_lines(char *filename);
     void init_history(envdata_t *environment);
     void free_history(history_t *history);
     void add_line_to_history(history_t *history, char *line);
-    char *history_get_line_from_offset(history_t *history, size_t offset);
+    char *history_get_line_from_offset(history_t *history, int offset);
     void show_history(history_t *history);
     void operate_on_previous_command(char *input, history_t *history);
     void operate_on_single_arg(char **input, history_t *history);
